hooks_ct: fix ub limits setting for upstream containers
Currently, vps_setup_res() have an explicit test for state != VPS_STARTING
before applying beancounter limits. This means that we can set limits without
further problems when the container is running, but will fail to do so when the
container is starting. In fact, hooks_vz have an explicit call for it in
its startup function.

We should do the same, and call our version of setlimits ourselves when the
container is coming up.

Signed-off-by: Glauber Costa <glommer@openvz.org>