hooks_ct: create devices inside container
Our devices were being created from the parent of container's init, because we
need to be still outside container context to do it. However, this creates
quite an annoyance, because those bind mounts will show up in the host
/proc/mounts.

Turns out, we don't really need to do it from the root side. We can do it from
the container side provided we do it before we chroot - and then the host side
fs is still visible.

The fact that we join a mount namespace will act to keep those mounts totally
private, and exempt us from cleaning it up.

Signed-off-by: Glauber Costa <glommer@openvz.org>