vzctl start: add pre-start dist script
We will need that infrastucture when running with Linux upstream, since some
support is very unlikely to ever land in the Kernel. This will be done by
running a vps script that will be run shortly before we call exec() on
/sbin/init.

To demonstrate what such script should do, this patch uses the script to
override the loginuid PAM module. This is called only for the distributions in
which I have verified the need and tested that the fixup works. We should look
into expanding this.

[kir@: fix subject, code indentation, and run on every kernel]

Signed-off-by: Glauber Costa <glommer@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>