ct_env_create_real(): don't use uninitialized var
In fact we do write() just to tell the other side it can proceed,
so it's not important if what we write is initialized or not.

But just to make Coverity happy, let's write 0.

Also, while at it, declare the variable closer to its use, and rename it
to something more neutral, since we don't mean any 'err' here.

Reported by Coverity, CID 1025649

Signed-off-by: Kir Kolyshkin <kir@openvz.org>