vzctl stop: don't kill CT if halt returned error
Sometimes CT's /sbin/halt returns an exit code of 1 but CT is still able
to shut down. In such case, we should still wait for CT to shut down,
and kill it forcibly only after a timeout.

This is a fix to commit b7baef97.

I was able to reproduce this behavior (halt returning 1) on a freshly
started CentOS 5 system with CTs started on boot. Not able to reproduce
it now though...

Signed-off-by: Kir Kolyshkin <kir@openvz.org>