Deny "unlimited" value for DISKSPACE/DISKINODES
Using a special value of "unlimited" was intended for user beancounters
parameters only. But since the same parsing functions are used for
diskspace and diskquota parameters, there was an ability to use
"unlimited" for these. This patch disables this.

Using "unlimited" for diskspace and diskinodes is just plain wrong.

For simfs/vzquota based container, the proper way of having unlimited
disk space is to disable vzquota, i.e. set DISK_QUOTA=no instead.

For ploop based container, 'unlimited' doesn't make sense at all, since
ploop image should be of some (limited) size. Using "unlimited", which
is translated to LONG_MAX, should create a ploop image with the size
of 2TB less 1 byte. This is impossible for V1 ploop format, and while
possible for ploop V2, is hardly what a user really wanted.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>