vzctl start: do fsck for ploop, add --skip-fsck
This patch adds
(1) Setting fsck flag for ploop_mount on CT start
(2) Option --skip-fsck for vzctl start command
    (together with man page and bash_completion changes)
(3) Using --skip-fsck from init.d/vz script

NOTE that while everything is in place, fsck flag is not actually
passed on to ploop library. It is  because the current version
of ploop library released (1.7.1) does not have it yet.
When ploop lib with fsck field in ploop_mount_param
will be released, we will need to uncomment the appropriate
assignment in vps_mount_image().

https://bugzilla.openvz.org/2615

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>