scripts/vps-netns_dev_add: rework config action
Commit 4fdbcb5 adds required parameter ACTION to vps-netns_dev_add
script, but ct_netdev_ctl() and ct_ip_ctl() do not set it. That
leads to the following bug:

	When trying to start a container with configured `IP_ADDRESS`, it fails
	with the subj message ("vps-netns_dev_add ERROR: Missing parameter:
	ACTION").

Instead of setting required parameter ACTION, let's just
set optional parameter SKIP_CREATE. That way we don't have
to set it in every place that calls that script.

While at it, document the parameter.

This is a fix to / slight rework of commit 4fdbcb5.

https://bugzilla.openvz.org/2637

Reported-by: hoverhell@gmail.com
Signed-off-by: Kir Kolyshkin <kir@openvz.org>