vzmigrate: improve invalid cmdline handing
This patch changes the way vzmigrate reacts to invalid command line.
Now, instead of showing a screen full of options, it just tells what's
wrong with the command line, and gives a hint on where to read about
the correct syntax, so the error is more clear.

Also, if --help, -h, --usage, or -? were given, or no options at all,
it shows usage as before.

Exit code logic is also fixed. When showing usage exit code is 0,
when showing error exit code is $MIG_ERR_USAGE.

Modify man page accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>