vzmigrate: add ability to ignore some checks
More preliminary checks were introduced recently to vzmigrate, such as
ipv6 module check (commit 9198710) and cpu capabilities compatibility
check (commit e22c459).

Say, for CPU flags, there is no easy way to find out whether or not
these capabilities are used by applications running in a particular
container -- but if they are, migration will fail. So, to be on the
safe side, vzmigrate is not proceeding with the migration.

Still, it makes sense to allow ignoring the result of these checks,
letting the migration process to continue. So -f, --nodeps option
is introduced, to specify all or some particular checks to ignore.

https://bugzilla.openvz.org/2643

Reported-by: Michal Gasewicz <genn@umk.pl>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>