vzlist -j: fix to work on RHEL5 kernel
vzlist --json is not working on RHEL5 kernel. The problem is caused by
the fact that in JSON mode all the parameters are listed by default, and
in RHEL5 kernel there is no way to get the value of cpus field from the
running container. vzlist should just ignore it, instead it fails.

Fix to skip cpus collection in case it's not available.

https://bugzilla.openvz.org/2661

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>