[build] fix for automake < 1.10.2
If someone wants to build vzctl from git on a RHEL5/CentOS 5 system
(or other similar system having automake < 1.10.2), they will discover
that pkglibexecdir is not set and scripts are installed to /scripts.

We could have said (in configure.ac)

	AM_INIT_AUTOMAKE([1.10.2

to require newer automake, but since we are still supporting RHEL5
let's workaround by setting pkglibexecdir manually.

NOTE we'll have to revert this one and do proper automake requirement
later.

This is an addition to commit d3c8b95.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>