check if container is actually running
Sometimes its possible that empty cgroup exists. For example
when we stop containers without --fast option. Such cgroups
could be reused and don't lead to any errors.
So let's assume that container is running when it has
at least one process on upstream kernel, or ve.state is
RUNNING on new vz kernel.