debian-add_ip.sh: ignore comments when looking for venet0
When starting a debian container, the /etc/vz/dists/scripts/debian-add_ip.sh
script will do some checks to automatically overwrite debian interfaces
config if the VM is in venet mode.

When we leave some old commented venet configuration in
/etc/network/interfaces, the script will detect it and overwrite it,
even if we don't want to because the network is in veth mode.

Steps to reproduce :
1. have a debian VM with working veth network
2. Add a comment line in /etc/network/interfaces : "# venet0"
3. Start or restart the debian VM

--> /etc/network/interfaces will be overwrited and network
broken in the VM.

https://bugzilla.openvz.org/2674