vzmigrate: remove a bashism
vzmigrate is unable to move CTs when the target host is a Debian
with /bin/dash as default /bin/sh shell (which is the default
on new Debian installs).

To make the vzmigrate script POSIX compatible, the (easy) fix is to
replace 'source' by '.'.

This is a fix to commit 0853972.

https://bugzilla.openvz.org/2682