vzctl start: always mount /dev/pts for upstream CT
Commit 5b8cd90, among other things, added devpts mount to container
start. For some reason it was done only if userns is available. In fact,
we need devpts inside a container in order to be able to enter it, so
remove the if (userns) condition.

This should fix vzctl enter on a recent Fedora 18-19 kernel
(3.9, the one that disables CONFIG_USERNS in favor of XFS).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>