vzmigrate: add / to paths for rsync
I think the way rsync is called in the vzmigrate script may
lead to data loss.

Steps to reproduce:
* have a VM on a source host
* add a trailing slash in VE_PRIVATE in the VM config (example:
* VE_PRIVATE="/vz/private/888/")
* have a target host with some no critical data in /vz/private
* vzmigrate your VM from source to target host (you may use
  --rsync="-vn" to have some debug and not to do the real rsync)

--> All /vz/private directory will be deleted on target host
(thanks to --delete rsync option)

This is because the effective rsync command will be :

rsync -aHAX --delete --numeric-ids -v \
	/vz/private/888/ root@target_host:/vz/private

I think it's better to call rsync with trailing slashes in directories,
to tell/force rsync to synchronize the directory contents.

https://bugzilla.openvz.org/2686