vzctl.spec: run vz-postinstall on a fresh install only
vz-postinstall changes some sysctl.conf parameters etc.
We do not need to run it every time we upgrade vzctl, but only
when installing on a fresh system (with no previous version of
vzctl installed). The thing is, if vzctl is installed, chances
are high system is already configured the way it should be.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>