load_ploop_lib(): load .so.1, try .so too
In ploop-1.9 we are introducing library versioning, so we have
libploop.so.1 and thus dlopen("libploop.so") will fail. Therefore,
we can't release ploop-1.9 before vzctl because it will break existing
systems. We can't release vzctl with dlopen("libploop.so.1") before
ploop-1.9 either for the same reason. Chicken and egg problem here.

Since we can't release vzctl and ploop at the same time, let's introduce
a way to try dlopen-ing libploop.so if libploop.so.1 failed, so we can
release vzctl and then later ploop. After both are released, in the next
release of vzctl we can remove this kludge and just require ploop-1.9.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>