ndsend: fix option field in sending packets
ndsend which is contained in vzctl package comes with an important bug.
According to RFC4861 #4.4 && #4.6.1 when sending an unsolicited Neighbor
Advertisement one should be using option type 2 (Target Link-Layer
Address). ndsend on the other hand uses option type 1 (Source Link-Layer
Address).

This makes RFC compliant devices not behaving according to what ndsend
is supposed to do, the packet is ignored.

https://bugzilla.openvz.org/2709
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=719095

Reported-by: Ola Lundqvist <ola@inguza.com>
Reported-by: George Kargiotakis <kargig@noc.grnet.gr>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>