vzctl.spec: don't bundle /var/lib/vz -> /vz symlink
For people who have /var/lib/vz on their systems, installation
of vzctl-4.5 (which includes commit df80be5) fails with the
following diagnostics:

  Updating   : vzctl-core-4.5-1.x86_64
 Error unpacking rpm package vzctl-core-4.5-1.x86_64
 error: unpacking of archive failed on file /var/lib/vz: cpio: rename

Apparently it was not a brilliant idea to bundle /var/lib/vz symlink
right to the rpm package. Instead, let's create a symlink in the
post-install script, if there's no /var/lib/vz.

https://bugzilla.openvz.org/2722

Reported-by: Pavel Snajdr <pavel.snajdr@relbit.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>