configure: fix libdir for Debian/Ubuntu case
In case we're building under Debian (or Ubuntu) on an x86_64, /usr/lib64
is not a good place for libraries. Instead, we should query
dpkg-architecture. Let's do that if dpkg-architecture is available,
otherwise fall back to /usr/lib64.

Note that this guessing still only happens if we are on x86_64
and libdir is not altered from the command line.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>