set_ublimit(): don't set unknown UBs to unlim
This basically reverts commit bfabefd. Apparently it was needed
for some very old 042* kernel, and was not taking into account
the fact that set_ublimit() is also called from the vzctl set
command line for a running container, when not all UBC parameters
are specified.

This fixes the following bug: when vzctl set --physpages is called
for a running container, it sets all the beancounters that were not
specified in command line to unlimited.

https://bugzilla.openvz.org/2760

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>