Don't modify sysctl.conf on installation
As suggested in the bug report linked below, installing vzctl rpm should
not modify /etc/sysctl.conf, plus there should be a way to totally
prevent it. Still, it's better to disable selinux when we install, i.e.
before a reboot.

So, what this patch does is:

1. Splits vz-postinstall functionality into 'sysctl' and 'selinux'
pieces.
2. Call vz-postinstall selinux from vzctl.rpm on a fresh install.
3. Call vz-postinstall sysctl from init.d/vz if SKIP_SYSCTL_SETUP
is not set in /etc/vz/vz.conf, and set it.
4. Document this in vz.conf(5) man page.

https://bugzilla.openvz.org/2735

Reported-by: Michael Renner <robe@amd.co.at>
Reported-by: Stefan Schlesinger <sts@ono.at>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>