vz-postinstall: don't add bridge params to sysctl.conf
This enhances commit a7aaea0. Since the default for these
net.bridge.bridge-nf-call-* parameters are 1, it doesn't make
sense to add those, we only need to change them to 1 if they
are there and set to 0 (RHEL case).

This should fix the following error message for distros that do not
load bridge by default:

	# sysctl -p
	error: "net.bridge.bridge-nf-call-ip6tables" is an unknown key
	error: "net.bridge.bridge-nf-call-iptables" is an unknown key

https://bugzilla.openvz.org/2743

Reported-by: poiuty <poiuty@lepus.su>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>