init.d/vzeventd: set reboot_event
The new RHEL6 kernel 042stab082.3, adds the following change:

	* ext4: send vzevent on error (PSBM-20496, PSBM-21010)

This change makes ext4 kernel module depend on vzevent, which is
loaded as a dependency. Therefore, it is loaded with defaults,
including reboot_event=0, and then init.d/vzeventd warns about it:

	/etc/init.d/vzeventd start
	Checking vzevent kernel module ...                         [FAILED]
	vzevent module should be loaded with reboot_event=1 paramet[WARNING]

Fortunately, there is an easy way to fix it -- just write "1" to
the relevant module parameter sysfs file. Then check it just to
be sure that write was successful.

Fix Gentoo script the same way.

https://bugzilla.openvz.org/2764

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>