vzubc: don't show unlimited ubcs, add -v
Now when most people use VSwap (or so I hope), there is no need
to show the beancounters that are set to unlimited. With this
patch, unlimited UBCs are not shown.

In non-VSwap mode, all UBCs are limited so they will still be
shown as before.

In VSwap mode, though, the output will be smaller thus simpler
to grasp.

Option -v|--verbose is added to switch vzubc to the old behavior,
i.e. show all UBCs. vzubc(8) man page is updated accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>