vzoversell: add
This simple script aims in finding out whether a system is oversold
in terms of RAM and swap. In other words, whether the sum of RAM+swap
limits for all container exceeds available RAM and swap.

Note that overselling RAM is OK since swap is used, while overselling
RAM+swap is not, since this could lead to OOM killing some processes.

NOTE this version is only usable for VSwap containers. For non-VSwap
containers it gives a rough guess on RAM limit based on privvmpages
limit (that is, privvmpages / 2), but this is very inaccurate, to say
at least. Go convert all your containers to VSwap already!

Signed-off-by: Kir Kolyshkin <kir@openvz.org>