vzctl set: require swap to be set for VSwap
In case we have a container with VSwap configuration either in
config file or supplied via the command line, we expect swap to
be set as well (again, either already in config file or via
command line). Otherwise a container would not be able to start
next time, due to check_ub().

So, let's just require swap to be set in this case.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>