vzctl start: don't fail if VE_ROOT does not exist
Since commit 78b5f53 we are not mounting CT if is_vps_mounted()
returns error. This is not the right thing to do since we fail
to start if VE_ROOT does not exist.

Fix is easy -- check for "if not mounted or unknown status".

https://bugzilla.openvz.org/2807

Reported-by: lopo@losys.eu
Signed-off-by: Kir Kolyshkin <kir@openvz.org>
(cherry picked from commit 6e512c2f13264ea37ac6660028a2eed5c37e6c76)