vzctl destroy: fix locking
vzctl destroy is performed asyncronously, i.e. we move the
private area to a temporary location, fork a background
process which then runs rm -rf which can take a lot of time.
Locking is introduced to avoid having multiple background
processes doing removal in parallel.

Commits b87d455 and then 32676b4f broke the locking logic.
Unlocking is supposed to be done on exit, so there's no reason
to do it explicitly.

https://bugzilla.openvz.org/2814

Reported-by: ton <donatas.abraitis@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>