debian-add_ip.sh: setup loopback device at least
After commit 346b5152 we are no longer touching /etc/network/interfaces
if there are no IPs to add or remove. While this is a good thing, it
could lead to a problem of not having loopback device setup, which is
apparently needed in Debian 7 to avoid 2 minutes boot delay.

This is one way to fix bug #2859.

https://bugzilla.openvz.org/2859

Signed-off-by: Kir Kolyshkin <kir@openvz.org>