vzctl start --wait: fix for non-standard Debian 7
Apparenly /etc/init/ directory existence doesn't mean that its contents
will be acted upon (i.e. it is not a guarantee that we are running
upstart).

Therefore, add_reach_runlevel_mark() should try to modify /etc/inittab
regardless of whether it placed a file in /etc/init/ (or /etc/event.d/).

Also, error out if system appears for not use upstart/systemd and
/etc/inittab is missing.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>