hooks_ct: remove non-working devpts mount
vzctl mounts devpts with the newinstance option. It doesn't work,
because CentOS umounts old devpts and tries to mount a new one.
Fedora mounts devtmpfs over /dev.

Instead of this hack, we need to add the newinstance option in
/etc/fstab or /etc/filesystems.

Another reason to remove it is that this configuration doesn't
supported by CRIU.

[kir@openvz.org: this in-container overmount is dealt with in prestart.sh,
but it is not working either, so applying this patch]

Signed-off-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>