prestart.sh: remove fixup_udev() completely
The whole point of fixup_udev() was to wait for container to be started
and then umount those mounts done from the inside of container, in order
to uncover the mounts previously done by vzctl.

Now, previous commit removes those mounts from vzctl code, as well
as umounts from this script. What is left in fixup_udev() is code
to wait for container to start. Remove that as well.

Cc: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>