vzctl snapshot-switch: add --skip_arpdetect option
If snapshot was done while CT was running, CT will be restored
from the dump file. While restoring, vps-net_add script is called
which, among the other things, calls arpsend to check if CT IPs are
used elsewhere. arpsend sends an ARP request for IP and waits for
up to one second for reply.

In case CT IPs are not changed from the current state, there is no
sense to run arpsend. So, add the option --skip_arpdetect to do that.

Note that since skip_arpdetect is set in cmd_p options, and we pass
g_p to vzctl_env_switch_snapshot, we need to explicitly copy the
skip_arpdetect option from cmd_p to g_p.

Note that since we re-read CT config on resume and use the new config
(if available), we need to explicitly copy the skip_arpdetect option
to the new param.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>