vzctl create, vzctl convert: honor diskinodes
For ploop layout, there is no way to change diskinodes because of
ext4 file system limitation. Therefore, there should be a way to
specify inodes on container creation (or conversion).

This patch does exactly that. On create or convert, diskinodes
(if set) are used to calculate disk space required to have
a requested number of inodes (based on the fact that by default
it's one inode per 16K disk space), and the filesystem of that
size is created. Then, if requested disk space is lower, fs
is resized down (with ballooning).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>