vzmigrate: call vzctl status with --quiet
Without --quiet, vzctl status output could be garbled if we run
vzmigrate with -vv (or higher verbosity) as in such case vzmigrate
calls vzctl with --verbose.

Nevertheless, it's a bug in vzctl, not vzmigrate, and it is fixed
by the previous commit 9bdd181. But we can't be sure it's fixed
on the remote side, so always use --quiet.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>