vzlist: don't complain about missing ploop-lib
Commit d1176e70037 introduced a way to silence ploop lib
for vzlist, since in vzlist we don't want to see anything
on stdout. The problem with that commit is it unconditionally
calls is_ploop_supported() which tries to load ploop library
and complains if it is missing, even then it is not needed:

	[root@ce6 ~]# vzlist -a
	Can't load ploop library: libploop.so: cannot open shared object file:
	No such file or directory
	Please install ploop packages!
	      CTID      NPROC STATUS    IP_ADDR         HOSTNAME

Surely, we want to use ploop functions only when it is really
necessary (i.e. CT layout is ploop and diskspace/diskquota
parameters are requested).

So, let's move ploop.set_verbose_level() to get_ve_ploop_info().
This way, we will

1. Not get a "Can't load ploop library" error when it's not needed
2. Get the above error when ploop-lib is needed but can't be loaded
   (ploop diskspace/diskinodes are requested, no ploop-lib found).

https://bugzilla.openvz.org/2952

Signed-off-by: Kir Kolyshkin <kir@openvz.org>