init.d/vz, vz.conf: use/expose VE_PARALLEL
Parallelism of VE start/stop in vz initscript was calculated
as number of CPUs * 4. It makes sense to give a user an ability
to change the number, so let's name it VE_PARALLEL and put to
global vz.conf file. If unset, it is still autocalculated the same way.

vz.conf(5) man page modified accordingly

Parameter name added into "well known" ones in src/lib/config.c
to avoid a warning from vzctl.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>