vzctl.spec: disable VE0 conntracks only if unused
Commit a191a462579ee (available in vzctl 4.7) added a post-installation
hook to disable VE0 conntracks for the sake of better venet performance.
While this is a good thing for many users, some are using NAT
for hardware node and this change requires manual editing of
configuration file and a reboot in order to fix things.

So, instead of just disabling VE0 conntracks, be a bit more clever
and try to figure out if those are used. If they are, do not disable
conntracks (but put a line to modprobe.d/openvz.conf anyway so the
next upgrade will just skip the hook).

https://bugzilla.openvz.org/2755

Reported-by: paparaciz <helpaz@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>