vzctl start,restore: add --skip-remount
By default, if a container is already mounted, it is remounted
right before start (or restore). This is a good behavior because
since the last time the container was mounted something might have
changed (container parameters or mount scripts, for example).

In some cases, though, it makes sense to mount a container and
then start or restore it without remounting, so this option might
be handy.

Also
* document it in vzctl(8) and vzctl --help
* add the option to bash-completion script

Signed-off-by: Kir Kolyshkin <kir@openvz.org>